+++ config.py
@@ -1,0 +1,2 @@
+# If there are any configuration values, put them here
+# The file run.py should read configs from here
